+++ changelog.md
@@ -1,4 +1,12 @@
 # Changelog
+## Next
+### Ethereum
+- added hyperledger besu
+- removed openzeppelin cli
+### Internal
+- renaming base script
+- moved install logic to network scripts
+
 ## 0.8.0 > 21 Sept 2020
 - [server] start_eth_project to create a base project in Ethereum
 - [server] ethereum.sh script, to install lighthouse, Prysm and a more complete dev environment